% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyze_inv_wishart_prior}
\alias{analyze_inv_wishart_prior}
\title{Analyze Inverse-Wishart Prior Parameters}
\usage{
analyze_inv_wishart_prior(nvar, nu = NULL, V = NULL)
}
\arguments{
\item{nvar}{Number of variables (dimension of the matrix).}

\item{nu}{Degrees of freedom (scalar > nvar - 1). Optional.}

\item{V}{A positive definite scale matrix (nvar x nvar). Optional.}
}
\value{
A list containing the expectation matrix 'E_X' and a vector
'Var_X_diag' with variances of the diagonal elements. Returns NA for
values if conditions on 'nu' are not met.
}
\description{
Calculates the expectation and variance of diagonal elements for an
Inverse-Wishart distribution W^-1(V, nu).
Uses defaults from \code{rhierMnlRwMixture} if \code{nu} or \code{V} are NULL.
}
\references{
\url{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}
}
