% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlBART-package.R
\docType{package}
\name{bayesm.HART-package}
\alias{bayesm.HART}
\alias{bayesm.HART-package}
\title{bayesm.HART: Hierarchical Logit Models with Bayesian Additive Regression Trees}
\description{
The \code{bayesm.HART} package provides a hierarchical Bayesian
machine learning approach for personalization. At its core, the package
implements a Metropolis-within-Gibbs sampler for a hierarchical logit
model with a Hierarchical Additive Regression Trees (HART) prior. HART is a
nonparametric prior that models the "representative consumer" as a flexible
sum-of-trees function of observed characteristics. This extends
conventional hierarchical models that are often limited to linear functions
of few characteristics.
}
\author{
\strong{Maintainer}: Thomas Wiemann \email{wiemann@uchicago.edu}

}
\keyword{internal}
