% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The expression to check.}

\item{b}{The default value if \code{a} is NULL.}
}
\value{
\code{a} if not NULL, otherwise \code{b}.
}
\description{
Provides a default value if the left-hand side expression evaluates to NULL.
Similar to \verb{rlang::\%||\%}.
}
\keyword{internal}
