% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{marginal_effects}
\alias{marginal_effects}
\alias{marginal_effects.rhierMnlRwMixture}
\alias{marginal_effects.rhierLinearMixture}
\alias{marginal_effects.rhierNegbinRw}
\title{Calculate Marginal Effects for Hierarchical Models}
\usage{
marginal_effects(object, ...)

\method{marginal_effects}{rhierMnlRwMixture}(object, z_values, Z, burn = 0, verbose = TRUE, ...)

\method{marginal_effects}{rhierLinearMixture}(object, z_values, Z, burn = 0, verbose = TRUE, ...)

\method{marginal_effects}{rhierNegbinRw}(object, z_values, Z, burn = 0, verbose = TRUE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The result depends on the method implementation.
}
\description{
Computes the posterior distribution of average marginal effects by
varying a target covariate over a grid.
}
\details{
This is a generic function. The main implementation for this package
is \code{marginal_effects.rhierMnlRwMixture}.
}
